refactor: 리뷰 반영, 값 객체가 값을 변경하지 않도록 수정

+++ src/main/java/lotto/domain/LottoMarket.java
@@ -11,7 +11,7 @@ public class LottoMarket {
     }
 
     private static long getLottoAmount(Money money) {
-        return money.divide(TICKET_PRICE);
+        return (long) money.divide(TICKET_PRICE);
     }
 
     private static List<LottoTicket> make(LottoGenerator lottoGenerator, long amount) {
